Reorganized app service into new file
